Fix Type Casting Bug in Scripts that plot the Renewal Events per Lithium Ion in each Layer

+++ scripts/bulk_and_walls/mdt/renewal_events_walls/plot_renewal_event_num_chain-length_dependence.py
@@ -303,7 +303,7 @@ ydata = [
     [[[] for sim in Sims.sims] for pkp_type in pk_pos_types]
     for dat_ix in range(n_data)
 ]
-n_events_per_refcmp_bulk = np.zeros(Sims.n_sims, dtype=np.uint32)
+n_events_per_refcmp_bulk = np.zeros(Sims.n_sims, dtype=np.float64)
 for sim_ix, Sim in enumerate(Sims.sims):
     # Get file that contains the renewal event information for the
     # corresponding bulk simulation.
